requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coderedcms==4.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.2
django-bootstrap5==24.2
django-debug-toolbar==4.4.6
django-eventtools==1.0.3
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-rosetta==0.10.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
future==1.0.0
geocoder==1.38.1
icalendar==5.0.13
idna==3.10
l18n==2021.3
laces==0.1.1
openpyxl==3.1.5
pillow==10.4.0
pillow_heif==0.20.0
polib==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
ratelim==0.1.6
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.2.3
wagtail==6.2.2
wagtail-cache==2.5.1
wagtail-localize==1.10
wagtail-seo==2.5.0
Willow==1.9.0
